% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grid.R
\name{data_grid}
\alias{data_grid}
\title{Create a reference grid.}
\usage{
data_grid(x, target = "all", length = 10, factors = "reference",
  numerics = "mean", preserve_range = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{x}{An object from which to contruct the reference grid.}

\item{target}{Can be "all" or list of characters indicating columns of interest. The remaining variables will be fixed.}

\item{length}{Length of numeric target variables.}

\item{factors}{Type of summary for factors. Can be "combination" or "reference".}

\item{numerics}{Type of summary for numerics Can be "combination", any function ("mean", "median", ...) or a value.}

\item{preserve_range}{In the case of combinations between numeric variables and factors, setting \code{preserve_range = TRUE} removes observerations where the value of the numeric variable is originally not present in the range of its factor level.}

\item{na.rm}{Remove NaNs.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a reference grid.
}
\examples{
newdata <- data_grid(iris, target = "Sepal.Length")
newdata <- data_grid(iris, target = "Sepal.Length", factors = "combinations")
newdata <- data_grid(iris, target = c("Sepal.Length", "Species"), length = 3)
newdata <- data_grid(iris, target = c("Sepal.Length", "Species"), numerics = 0)
}
