% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grid.R
\name{data_grid.vector}
\alias{data_grid.vector}
\title{Create a reference grid.}
\usage{
\method{data_grid}{vector}(x, target = "all", length = 10, ...)
}
\arguments{
\item{x}{An object from which to contruct the reference grid.}

\item{target}{Can be "all" or list of characters indicating columns of interest. The remaining variables will be fixed.}

\item{length}{Length of numeric target variables.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a reference grid.
}
