% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_restoretype.R
\name{data_restoretype}
\alias{data_restoretype}
\title{Restore the type of columns according to a reference data.frame}
\usage{
data_restoretype(data, reference_data = NULL)
}
\arguments{
\item{data}{The data.}

\item{reference_data}{A reference dataframe from which to find the correct
column types.}
}
\description{
Restore the type of columns according to a reference data.frame
}
\examples{
library(modelbased)

x <- data.frame(
  Sepal.Length = c("1", "3", "2"),
  Species = c("setosa", "versicolor", "setosa")
)
fixed <- data_restoretype(x, reference_data = iris)
summary(fixed)
}
