% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_nonlinear.R
\name{describe_nonlinear}
\alias{describe_nonlinear}
\alias{estimate_smooth}
\title{Describe the smooth term (for GAMs) or non-linear predictors}
\usage{
describe_nonlinear(data, y = NULL, x = NULL, ...)

estimate_smooth(data, y = NULL, x = NULL, ...)
}
\arguments{
\item{...}{Other arguments passed for instance to \code{\link{visualisation_matrix}}.}

\item{nonlinear}{A character indicating the name of the "smooth" term}

\item{smooth}{Same as \code{nonlinear} (added for compatibility reasons with old API).}
}
\value{
A dataframe of linear description of non-linear terms.
}
\description{
This function summarises the smooth term trend in terms of linear segments.
Using the approximative derivative, it separates a non-linear vector into
quasi-linear segments (in which the trend is either positive or negative).
Each of this segment its characterized by its beginning, end, size (in
proportion, relative to the total size) trend (the linear regression
coefficient) and linearity (the R2 of the linear regression).
}
\examples{
library(modelbased)

# Create data
data <- data.frame(x = rnorm(200))
data$y <- data$x^2 + rnorm(200, 0, 0.5)

model <- lm(y ~ poly(x, 2), data = iris[1:50, ])
link_data <- estimate_relation(model, length = 100, preserve_range = FALSE)

}
