% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{.uniroot.all}
\alias{.uniroot.all}
\title{Copied from rootSolve package}
\usage{
.uniroot.all(f, interval, lower = min(interval), upper = max(interval),
  tol = .Machine$double.eps^0.2, maxiter = 1000, n = 100, ...)
}
\description{
Copied from rootSolve package
}
\keyword{internal}
