% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efc}
\alias{efc}
\title{Sample dataset from the EFC Survey}
\description{
Selected variables from the EUROFAMCARE survey. Useful when
testing on "real-life" data sets, including random missing values. This
data set also has value and variable label attributes.
}
\keyword{data}
