% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_individual.R
\name{estimate_individual}
\alias{estimate_individual}
\title{Group-specific parameters of mixed models}
\usage{
estimate_individual(model, indices = c("Coefficient", "SE"), ...)
}
\arguments{
\item{model}{A mixed model.}

\item{indices}{A list containing the indices to extract.}

\item{...}{Other arguments passed to or from other methods.}
}
\description{
Extract parameters of each individual group in the context of mixed models.
}
\examples{
library(modelbased)

if (require("lme4")) {
  data <- lme4::sleepstudy

  # Random intercept
  model <- lmer(Reaction ~ Days + (1 | Subject), data = data)
  random <- estimate_individual(model)
  head(random)
  summary(random)

  # Random slope and intercept
  model <- lmer(Reaction ~ Days + (1 + Days | Subject), data = data)
  random <- estimate_individual(model)
  head(random)

  # Multiple random groups
  model <- lmer(mpg ~ wt + (1 | gear) + (1 | carb), data = mtcars)
  random <- estimate_individual(model)
  head(random)

  # Nested random factors
  set.seed(33)
  data$grp <- sample(letters[1:5], size = 180, replace = TRUE)
  data$subgrp <- NA
  for (i in letters[1:5]) {
    filter_group <- data$grp == i
    data$subgrp[filter_group] <- sample(LETTERS, size = sum(filter_group), replace = TRUE)
  }
  model <- lmer(Reaction ~ Days + (1 | grp / subgrp) + (1 | Subject), data = data)
  # estimate_individual(model)
}
}
