% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_random.R, R/reshape_random.R
\name{estimate_random}
\alias{estimate_random}
\alias{reshape_random}
\title{Group-specific parameters of mixed models random effects}
\usage{
estimate_random(model, ...)

reshape_random(x, indices = "all", ...)
}
\arguments{
\item{model}{A mixed model with random effects.}

\item{...}{Other arguments passed to or from other methods.}

\item{x}{The output of \code{estimate_random()}.}

\item{indices}{A list containing the indices to extract.}
}
\description{
Extract random parameters of each individual group in the context of mixed models. Can be reshaped to be of the same dimensions as the original data, which can be useful to add the random effects to the original data.
}
\examples{
if (require("lme4")) {

# Random intercept ------------
model <- lmer(mpg ~ hp + (1 | carb), data = mtcars)
random <- estimate_random(model)
random

# Reshape to wide data so that it matches the original dataframe...
reshaped <- reshape_random(random, indices = c("Coefficient", "SE"))

# ... and can be easily combined
alldata <- cbind(mtcars, reshaped)

# Use summary() to remove duplicated rows
summary(reshaped)
}
}
