% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_response.R
\name{estimate_response}
\alias{estimate_response}
\title{Generates predictions}
\usage{
estimate_response(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_response.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
\examples{
\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Width ~ Species * Petal.Length, data=iris)
estimate_response(model)

model <- stan_glmer(Sepal.Width ~ Petal.Length + (1|Species), data=iris)
estimate_response(model)
}


}
