% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_response.R
\name{estimate_response.stanreg}
\alias{estimate_response.stanreg}
\alias{estimate_fit.stanreg}
\alias{estimate_response.data.frame}
\alias{estimate_fit.data.frame}
\title{Generates predictions for Bayesian models}
\usage{
\method{estimate_response}{stanreg}(model, data = NULL,
  predict = "response", ci = 0.9, estimate = "median",
  transform = "response", keep_draws = FALSE, draws = NULL,
  seed = NULL, random = FALSE, length = 10, ...)

\method{estimate_fit}{stanreg}(model, data = "grid", predict = "link",
  ci = 0.9, estimate = "median", transform = "response",
  keep_draws = FALSE, draws = NULL, seed = NULL, random = FALSE,
  ...)

\method{estimate_response}{data.frame}(model, data = NULL,
  predict = "response", ci = 0.9, estimate = "median",
  transform = "response", keep_draws = FALSE, draws = NULL,
  seed = NULL, random = FALSE, ...)

\method{estimate_fit}{data.frame}(model, data = "grid",
  predict = "link", ci = 0.9, estimate = "median",
  transform = "response", keep_draws = FALSE, draws = NULL,
  seed = NULL, random = FALSE, ...)
}
\arguments{
\item{model}{Bayesian model.}

\item{data}{A data frame with model's predictors to estimate the response. If NULL, the model matrix is used. If "grid", the model's data is re-used.}

\item{predict}{Can be "response" (default) or "link". The former predicts the the outcome per se, while the latter predicts the link function (i.e., the regression "line"), equivalent to estimating the \code{fit}. In other words, \code{estimate_response(model, predict="link")} is equivalent to \code{estimate_fit(model)}.}

\item{ci}{Credible Interval (CI) level. Default to 0.90 (90\%).}

\item{estimate}{The \href{https://easystats.github.io/bayestestR/articles/2_IndicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{transform}{Can be "none", "response", "mu", "unlink", "log". "none" (default for contrasts) will leave the values on scale of the linear predictors. "response" (default for means) will transform them on scale of the response variable. Thus for a logistic model the default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities.}

\item{keep_draws}{If FALSE, will summarise the posterior the obtained distributions. If TRUE, will keep all prediction iterations (draws).}

\item{draws}{An integer indicating the number of draws to return. The default and maximum number of draws is the size of the posterior sample contained in the model.}

\item{seed}{An optional seed to use.}

\item{random}{Should it take the random effects into account? Can be \code{TRUE}, \code{FALSE} or a formula indicating which group-level parameters to condition on when making predictions. The data argument may include new levels of the grouping factors that were specified when the model was estimated, in which case the resulting posterior predictions marginalize over the relevant variables (see \code{posterior_predict.stanreg}).}

\item{length}{Length of numeric target variables that applies to \link{data_grid}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Generates predictions for Bayesian models
}
