% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_slopes.R
\name{estimate_slopes.stanreg}
\alias{estimate_slopes.stanreg}
\title{Estimate the marginal slope of a numeric predictor (over different factor levels)}
\usage{
\method{estimate_slopes}{stanreg}(
  model,
  trend = NULL,
  levels = NULL,
  transform = "response",
  standardize = TRUE,
  standardize_robust = FALSE,
  ci = 0.95,
  centrality = "median",
  ci_method = "hdi",
  test = c("pd", "rope"),
  rope_range = "default",
  rope_ci = 1,
  ...
)
}
\arguments{
\item{trend}{A character vector indicating the name of the numeric variable
for which to compute the slopes.}

\item{levels}{A character vector indicating the variables over which the
slope will be computed. If NULL (default), it will select all the remaining
predictors.}
}
\description{
Estimate the marginal slope of a numeric predictor (over different factor levels)
}
\examples{
library(modelbased)
\donttest{
if (require("rstanarm")) {
  model <- stan_glm(Sepal.Width ~ Species * Petal.Length, data = iris)
  estimate_slopes(model)
}
}
}
