% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_smooth.R
\name{estimate_smooth}
\alias{estimate_smooth}
\title{Describe the smooth term (for GAMs) or non-linear predictors}
\usage{
estimate_smooth(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
This function summarise the smooth term trend in terms of linear segments. See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_smooth.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
