% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{find_inversions}
\alias{find_inversions}
\title{Find points of inversion}
\usage{
find_inversions(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Vector of inversion points.
}
\description{
Find points of inversion of a curve.
}
\examples{
x <- sin(seq(0, 4 * pi, length.out = 100))
plot(x, type = "b")
find_inversions(x)
}
