% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_marginaleffects.R
\name{model_marginaleffects}
\alias{model_marginaleffects}
\title{Easy marginaleffects}
\usage{
model_marginaleffects(model, trend = NULL, at = NULL, fixed = NULL, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{trend}{A character indicating the name of the variable
for which to compute the slopes.}

\item{at}{The predictor variable(s) \emph{at} which to evaluate the desired effect
/ mean / contrasts. Other predictors of the model that are not included
here will be collapsed and "averaged" over (the effect will be estimated
across them).}

\item{fixed}{A character vector indicating the names of the predictors to be
"fixed" (i.e., maintained), so that the estimation is made at these values.}

\item{...}{Other arguments passed for instance to \code{\link[=visualisation_matrix]{visualisation_matrix()}}.}
}
\description{
Modelbased-like API to create \pkg{marginaleffects} objects. This is Work-in-progress.
}
\examples{
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

model_marginaleffects(model, trend = "Petal.Length", at = "Species")
model_marginaleffects(model, trend = "Petal.Length", at = "Petal.Length")
model_marginaleffects(model, trend = "Petal.Length", at = c("Species", "Petal.Length"))


}
