% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{options}
\name{modelbased-options}
\alias{modelbased-options}
\title{Global options from the modelbased package}
\description{
Global options from the modelbased package
}
\section{Global Options to Customize Estimation of Marginal Means}{

Columns and table layout can be customized using \code{options()}:

\strong{For calculating marginal means}
\itemize{
\item \code{modelbased_backend}: \verb{options(modelbased_backend = <string>)} will set a
default value for the \code{backend} argument and can be used to set the package
used by default to calculate marginal means. Can be \code{"marginalmeans"} or
\code{"emmeans"}.
\item \code{modelbased_estimate}: \verb{options(modelbased_estimate = <string>)} will
set a default value for the \code{estimate} argument, which modulates the type
of target population predictions refer to.
}

\strong{For printing}
\itemize{
\item \code{modelbased_select}: \verb{options(modelbased_select = <string>)} will set a
default value for the \code{select} argument and can be used to define a custom
default layout for printing.
\item \code{modelbased_include_grid}: \code{options(modelbased_include_grid = TRUE)} will
set a default value for the \code{include_grid} argument and can be used to
include data grids in the output by default or not.
\item \code{modelbased_full_labels}: \code{options(modelbased_full_labels = FALSE)} will
remove redundant (duplicated) labels from rows.
}

\strong{For plotting}
\itemize{
\item \code{modelbased_join_dots}: \verb{options(modelbased_join_dots = <logical>)} will
set a default value for the \code{join_dots}.
\item \code{modelbased_numeric_as_discrete}: \verb{options(modelbased_numeric_as_discrete = <number>)}
will set a default value for the \code{modelbased_numeric_as_discrete} argument.
Can also be \code{FALSE}.
}
}

