% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbased-package.R
\docType{package}
\name{modelbased-package}
\alias{modelbased-package}
\alias{modelbased}
\title{modelbased: Estimation of Model-Based Predictions, Contrasts and Means}
\description{
\code{modelbased} is a package helping with model-based estimations, to
easily compute of marginal means, contrast analysis and model predictions.
}
\details{
\code{modelbased}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://easystats.github.io/modelbased/}
  \item Report bugs at \url{https://github.com/easystats/modelbased/issues}
}

}
\author{
\strong{Maintainer}: Dominique Makowski \email{dom.makowski@gmail.com} (\href{https://orcid.org/0000-0001-5375-9967}{ORCID}) (@Dom_Makowski)

Authors:
\itemize{
  \item Daniel Lüdecke \email{d.luedecke@uke.de} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID}) (@strengejacke)
  \item Mattan S. Ben-Shachar \email{matanshm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-4287-4801}{ORCID}) (@mattansb)
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets)
}

}
\keyword{internal}
