% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.estimate_contrasts}
\alias{print.estimate_contrasts}
\title{Printing modelbased-objects}
\usage{
\method{print}{estimate_contrasts}(
  x,
  select = getOption("modelbased_select", NULL),
  include_grid = getOption("modelbased_include_grid", FALSE),
  full_labels = TRUE,
  ...
)
}
\arguments{
\item{x}{An object returned by the different \verb{estimate_*()} functions.}

\item{select}{Character string, indicating which columns should be included
in the table.}

\item{include_grid}{Logical, if \code{TRUE}, the data grid is included in the
table output. Only applies to prediction-functions like \code{estimate_relation()}
or \code{estimate_link()}.}

\item{full_labels}{Logical, if \code{TRUE} (default), all labels for focal terms
are shown. If \code{FALSE}, redundant (duplicated) label rows are removed.}

\item{...}{Arguments passed to \code{insight::format_table()} or
\code{insight::export_table()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
\code{print()} method for \strong{modelbased} objects. Can be used to tweak the output
of tables.
}
\note{
Use \code{print_html()} and \code{print_md()} to create tables in HTML or
markdown format, respectively.
}
\section{Global Options to Customize Tables when Printing}{

Columns and table layout can be customized using \code{options()}:
\itemize{
\item \code{modelbased_select}: \verb{options(modelbased_select = <string>)} will set a
default value for the \code{select} argument and can be used to define a custom
default layout for printing.
\item \code{modelbased_include_grid}: \code{options(modelbased_include_grid = TRUE)} will
set a default value for the \code{include_grid} argument and can be used to
include data grids in the output by default or not.
}
}

\examples{
\dontshow{if (insight::check_if_installed("marginaleffects", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model <- lm(Petal.Length ~ Species, data = iris)
out <- estimate_means(model, "Species")

# default
print(out)

# smaller set of columns
print(out, select = "basic")
\dontshow{\}) # examplesIf}
}
