% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puppy_love}
\alias{puppy_love}
\title{More puppy therapy data}
\description{
Fictitious data related to whether puppy therapy works when you
adjust for a person’s love of puppies, taken from the \code{{discovr}} package
(Field 2025)
}
\details{
Following variables are included in the dataset:
\itemize{
\item \code{id}: Participant id
\item \code{dose}: Treatment group to which the participant was randomly assigned (No
puppies (control), 15 minutes of puppy therapy, 30 minutes of puppy
therapy)
\item \code{happiness}: Self-reported happiness from 0 (as unhappy as I can possibly
imagine being) to 10 (as happy as I can possibly imagine being)
\item \code{puppy_love}: Self-reported love of puppies from 0 (I am a weird person who
hates puppies, please be deeply suspicious of me) to 7 (puppies are the
best thing ever, one day I might marry one)
}

For further details, see \code{?discovr::puppy_love}.
}
\references{
Field, A. P. (2025). Discovering statistics using R and RStudio
(2nd ed.). London: Sage.
}
\keyword{data}
