% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_response.R
\name{reshape_draws}
\alias{reshape_draws}
\title{Reshape Estimations with Draws to long format}
\usage{
reshape_draws(estimates)
}
\arguments{
\item{estimates}{Estimates with posterior draws.}
}
\description{
Reshape Estimations with Draws to long format
}
\examples{
\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Width ~ Species * Petal.Length, data=iris)
estimates <- estimate_response(model, keep_draws=TRUE, draws=200)
reshape_draws(estimates)
}
}
