% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_recipe.R,
%   R/visualisation_recipe.estimate_predicted.R
\name{visualisation_recipe}
\alias{visualisation_recipe}
\alias{visualisation_recipe.estimate_predicted}
\title{Visualization helper}
\usage{
visualisation_recipe(x, ...)

\method{visualisation_recipe}{estimate_predicted}(
  x,
  show_points = TRUE,
  points = NULL,
  line = NULL,
  ribbon = NULL,
  labs = NULL,
  ...
)
}
\arguments{
\item{x}{Something.}

\item{...}{Other arguments passed to other functions.}
}
\description{
Visualization helper
}
\examples{
library(modelbased)

# Linear Models
# ----------------
# Simple
model <- lm(mpg ~ wt, data = mtcars)
x <- estimate_relation(model)
layers <- visualisation_recipe(x)
layers

if (require("ggplot2")) {
  ggplot() +
    geom_line(data = layers$l1$data,
              aes_string(x = layers$l1$x, y = layers$l1$y)) +
    geom_ribbon(data = layers$l2$data, alpha = layers$l2$alpha,
                aes_string(x = layers$l2$x, y = layers$l2$y,
                           ymin = layers$l2$ymin, ymax = layers$l2$ymax)) +
    labs(x = layers$l3$x, y = layers$l3$y, title = layers$l3$title)
}

# 2-ways interaction between numeric
model <- lm(mpg ~ wt * am, data = mtcars)
x <- estimate_relation(model)
layers <- visualisation_recipe(x)
layers

x <- estimate_response(model)
# visualisation_recipe(x)
}
