% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.estimate_contrasts}
\alias{as.data.frame.estimate_contrasts}
\title{Converting modelbased-objects into raw data frames}
\usage{
\method{as.data.frame}{estimate_contrasts}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE,
  use_responsename = FALSE
)
}
\arguments{
\item{x}{An object returned by the different \verb{estimate_*()} functions.}

\item{...}{Arguments passed to \code{as.data.frame()}.}
}
\value{
A data frame.
}
\description{
\code{as.data.frame()} method for \strong{modelbased} objects. Can be used to return
a "raw" data frame without attributes and standardized column names.
}
\examples{
\dontshow{if (insight::check_if_installed("marginaleffects", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model <- lm(Petal.Length ~ Species, data = iris)
out <- estimate_means(model, "Species")

# default
print(out)

as.data.frame(out)

as.data.frame(out, use_responsename = TRUE)
\dontshow{\}) # examplesIf}
}
