% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_contrasts.R
\name{estimate_contrasts}
\alias{estimate_contrasts}
\title{Estimate contrasts between factor levels}
\usage{
estimate_contrasts(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Contrast analysis. See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_contrasts.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
