% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_predicted.R
\name{estimate_link}
\alias{estimate_link}
\alias{estimate_relation}
\alias{estimate_expectation}
\alias{estimate_response}
\alias{estimate_prediction}
\title{Model-based response estimates and uncertainty}
\usage{
estimate_link(model, data = "grid", ci = 0.95, keep_iterations = FALSE, ...)

estimate_relation(
  model,
  data = "grid",
  ci = 0.95,
  keep_iterations = FALSE,
  ...
)

estimate_expectation(
  model,
  data = NULL,
  ci = 0.95,
  keep_iterations = FALSE,
  ...
)

estimate_response(...)

estimate_prediction(
  model,
  data = NULL,
  ci = 0.95,
  keep_iterations = FALSE,
  ...
)
}
\arguments{
\item{model}{A statistical model.}

\item{data}{A data frame with model's predictors to estimate the response. If
\code{NULL}, the model's data is used. If "grid", the model matrix is obtained
(through \code{\link[=visualisation_matrix]{visualisation_matrix()}}).}

\item{ci}{Confidence Interval (CI) level. Default to \code{0.95} (\verb{95\%}).}

\item{keep_iterations}{If \code{TRUE}, will keep all iterations (draws) of
bootstrapped or Bayesian models. They will be added as additional columns
named \verb{iter_1, iter_2, ...}. You can reshape them to a long format by
running \code{\link[bayestestR:reshape_iterations]{reshape_iterations()}}.}

\item{...}{You can add all the additional control arguments from \code{\link[=visualisation_matrix]{visualisation_matrix()}} (used when \code{data = "grid"}) and \code{\link[insight:get_predicted]{insight::get_predicted()}}.}
}
\value{
A data frame of predicted values and uncertainty intervals, with class \code{"estimate_predicted"}. Methods for \code{\link[=visualisation_recipe.estimate_predicted]{visualisation_recipe()}} and \code{\link[=visualisation_recipe.estimate_predicted]{plot()}} are available.
}
\description{
After fitting a model, it is useful generate model-based estimates of the response variables for different combinations of predictor values.
Such estimates can be used to make inferences about relationships between variables and to make predictions about individual cases.
\cr\cr
Model-based estimates and uncertainty can be generated for both the conditional average response values (the regression line) and for predictions about individual cases.
For generalized linear models, model-based estimates and uncertainty can also be reported on the original response scale or on the link scale.
See below for details.
}
\details{

}
\note{
These functions are built on top of \code{\link[insight:get_predicted]{insight::get_predicted()}} and correspond to different specifications of its parameters. It may be useful to read its \href{https://easystats.github.io/insight/reference/get_predicted.html}{documentation}, in particular the description of the \code{predict} argument for additional details on the difference between expected vs. predicted values and link vs. response scales.
\cr\cr

Additional control parameters can be used to control results from \code{\link[=visualisation_matrix]{visualisation_matrix()}} (when \code{data = "grid"}) and from \code{\link[insight:get_predicted]{insight::get_predicted()}} (the function used internally to comptue predictions).
\cr\cr

For plotting, check the examples in \code{\link[=visualisation_recipe]{visualisation_recipe()}}. Also check out the \href{https://easystats.github.io/modelbased/articles/}{Vignettes} and \href{https://easystats.github.io/modelbased/index.html#features}{README examples} for various examples, tutorials and usecases.
}
\section{Expected (average) values versus individual case predictions (forecasts)}{


The most important way that various types of predictions differ is in terms of what quantity is being estimated and the meaning of the uncertainty intervals.
\itemize{
\item \strong{Expected} values refer the the fitted regression line---the estimated \strong{average} response value (i.e., the "expectation") for individuals with a specific set of predictor values. For example, in a linear model \emph{y_i} = 2 + 3\emph{x_i} + \emph{e_i}, the estimated average \emph{y} for individuals with \emph{x} = 2 is 8. For expected values, uncertainty intervals refer to uncertainty in the estimated \strong{conditional average} (where might the true regression line actually fall)? Uncertainty intervals for expected values are also called "confidence intervals". Expected values and their uncertainty intervals are useful for describing the general relationship between variables and for describing how precisely a model has been estimated.
\item \strong{Predicted} values refer to forecasts or predictions for \strong{individual cases}. Predicted values are also called "posterior predictions". For predicted values, uncertainty intervals refer to uncertainty in the \strong{individual response values for each case} (where might any single case actually fall)? Uncertainty intervals for predicted values are also called "prediction intervals" and "posterior predictive intervals". Predicted values and their uncertainty intervals are useful for forecasting the range of values that might be observed in new data or for making decisions about individual cases.
}
}

\section{Response metric}{


Various types of predictions also differ in the metric or scale the predictions are expressed in.
\itemize{
\item The \strong{link scale} refers to scale of the fitted regression line. For linear models, this is the original response variable scale. However, for generalized linear models, it is the scale of the response after transformation by the link function (e.g., log-odds for logit binomial models, log-counts for log-linked Poisson models, log-probability for log-linked beta models).
\item The \strong{response scale} refers to the original scale of the response variable (i.e., without any link function transformation). For linear models, this is the original response variable scale. However, for generalized linear models, expected values on the link scale are back-transformed to the original response variable metric (e.g., predicted probabilities for binomial models, counts for Poisson models, predicted probabilities for beta models).
}
}

\section{\emph{modelbased} functions for estimating predicted values and uncertainty}{


\emph{modelbased} provides 4 functions for generating different types of predictions and their uncertainty from models:
\itemize{
\item \strong{\code{estimate_link()}}: Generates \strong{expected values} (conditional average) on the \strong{link scale}. The uncertainty interval is a \emph{confidence interval}. By default, values are estimated using a reference grid spanning the observed range of predictor values (see \code{\link[=visualisation_matrix]{visualisation_matrix()}}).
\item \strong{\code{estimate_expectation()}}: Generates \strong{expected values} (conditional average) on the \strong{response scale}. The uncertainty interval is a \emph{confidence interval}. By default, values are estimated using the data used to fit the model.
\item \strong{\code{estimate_prediction()}}: Generates \strong{predicted values} (for individual cases) on the \strong{response scale}. The uncertainty interval is a \emph{prediction interval}. By default, values are estimated using the data used to fit the model.
\item \strong{\code{estimate_relation()}}: Like \code{estimate_expectation()}. Generates \strong{expected values} (conditional average) on the \strong{response scale}. The uncertainty interval is a \emph{confidence interval}. By default, values are estimated using a reference grid spanning the observed range of predictor values (see \code{\link[=visualisation_matrix]{visualisation_matrix()}}).
}

\code{estimate_response()} is a deprecated alias for \code{estimate_expectation()}.
}

\section{Data for predictions}{


If the \code{data = NULL}, values are estimated using the data used to fit the model. If \code{data = "grid"}, values are estimated a reference grid spanning the observed range of predictor values with \code{\link[=visualisation_matrix]{visualisation_matrix()}}. This can be useful for model visualization. The number of predictor values used for each variable can be controlled with the \code{length} argument. \code{data} can also be a data frame containing columns with names matching the model frame (see \code{\link[insight:get_data]{insight::get_data()}}). This can be used to generate model predictions for specific combinations of predictor values.
}

\examples{
library(modelbased)

# Linear Models
model <- lm(mpg ~ wt, data = mtcars)

# Get predicted and prediction interval (see insight::get_predicted)
estimate_response(model)

# Get expected values with confidence interval
pred <- estimate_relation(model)
pred

# Visualisation (see visualisation_recipe())
plot(pred)

# Standardize predictions
pred <- estimate_relation(lm(mpg ~ wt + am, data = mtcars))
z <- effectsize::standardize(pred, include_response = FALSE)
z
effectsize::unstandardize(z, include_response = FALSE)

# Logistic Models
model <- glm(vs ~ wt, data = mtcars, family = "binomial")
estimate_response(model)
estimate_relation(model)

# Mixed models
if (require("lme4")) {
  model <- lmer(mpg ~ wt + (1 | gear), data = mtcars)
  estimate_response(model)
  estimate_relation(model)
}

# Bayesian models
\donttest{
if (require("rstanarm")) {
  model <- rstanarm::stan_glm(mpg ~ wt, data = mtcars, refresh = 0, iter = 200)
  estimate_response(model)
  estimate_relation(model)
}
}
}
