% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_means.R
\name{estimate_means}
\alias{estimate_means}
\title{Estimate average value of response at each factor levels}
\usage{
estimate_means(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_means.stanreg]{Bayesian models (rstanarm and brms)}}
 }
}
