% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_slopes.R
\name{estimate_slopes}
\alias{estimate_slopes}
\alias{estimate_slopes.glmmTMB}
\title{Estimate the slopes of a numeric predictor (over different factor levels)}
\usage{
estimate_slopes(
  model,
  trend = NULL,
  levels = NULL,
  transform = "response",
  standardize = TRUE,
  standardize_robust = FALSE,
  ci = 0.95,
  ...
)

\method{estimate_slopes}{glmmTMB}(
  model,
  trend = NULL,
  levels = NULL,
  transform = "response",
  standardize = TRUE,
  standardize_robust = FALSE,
  ci = 0.95,
  component = c("conditional", "zero_inflated", "zi"),
  ...
)
}
\arguments{
\item{trend}{A character vector indicating the name of the numeric variable
for which to compute the slopes.}

\item{levels}{A character vector indicating the variables over which the
slope will be computed. If NULL (default), it will select all the remaining
predictors.}

\item{component}{A character vector indicating the model component for which
estimation is requested. Only applies to models from \pkg{glmmTMB}. Use
\code{"conditional"} for the count-model or \code{"zero_inflate"} or
\code{"zi"} for the zero-inflation model.}
}
\value{
A data frame of slopes.
}
\description{
See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_slopes.lm]{Frequentist models}}
 \item{\link[=estimate_slopes.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
