% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_slopes.R
\name{estimate_slopes}
\alias{estimate_slopes}
\title{Estimate the slopes of a numeric predictor (over different factor levels)}
\usage{
estimate_slopes(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_slopes.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
