% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_slopes.R
\name{estimate_slopes}
\alias{estimate_slopes}
\title{Estimate Marginal Effects}
\usage{
estimate_slopes(model, trend = NULL, levels = NULL, ci = 0.95, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{trend}{A character vector indicating the name of the numeric variable
for which to compute the slopes.}

\item{levels}{A character vector or formula specifying the names of the
predicting factors over which to estimate the means, contrasts or slopes.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{...}{Other arguments passed for instance to \code{\link{visualisation_matrix}}.}
}
\value{
A data.frame.
}
\description{
Estimate the slopes (i.e., the coefficient) of a predictor over different factor levels. See also other
related functions such as \code{\link{estimate_contrasts}} and \code{\link{estimate_means}}.
}
\examples{
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)
slopes <- estimate_slopes(model, trend = "Petal.Length")
slopes
effectsize::standardize(slopes)
\dontrun{
if (require("rstanarm")) {
  model <- stan_glm(Sepal.Width ~ Species * Petal.Length, data = iris, refresh = 0)
  estimate_slopes(model)
}
}

}
