% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_slopes.R
\name{estimate_slopes}
\alias{estimate_slopes}
\title{Estimate Marginal Effects}
\usage{
estimate_slopes(
  model,
  trend = NULL,
  levels = NULL,
  modulate = NULL,
  ci = 0.95,
  ...
)
}
\arguments{
\item{model}{A statistical model.}

\item{trend}{A character vector indicating the name of the numeric variable
for which to compute the slopes.}

\item{levels}{A character vector or formula specifying the names of the
factors over which levels to estimate the means, contrasts or slopes. If
\code{NULL} (default), will automatically try to find it.}

\item{modulate}{A character vector indicating the names of a numeric variable
along which the means or the contrasts will be estimated. Other arguments
from \code{\link{visualisation_matrix}}, such as \code{length} to adjust the
number of data points.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{...}{Other arguments passed for instance to \code{\link{visualisation_matrix}}.}
}
\value{
A data.frame.
}
\description{
Estimate the slopes (i.e., the coefficient) of a predictor over different factor levels or alongside a numeric variable. See also other related functions such as \code{\link{estimate_contrasts}} and \code{\link{estimate_means}}.
}
\examples{
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)
slopes <- estimate_slopes(model, trend = "Petal.Length", levels = "Species")
slopes
plot(slopes)
effectsize::standardize(slopes)

if (require("mgcv")) {
  model <- mgcv::gam(Sepal.Width ~ s(Petal.Length), data = iris)
  slopes <- estimate_slopes(model, modulate = "Petal.Length", length = 30)
  summary(slopes)
  plot(slopes)
}
}
