% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_smooth.R
\name{estimate_smooth}
\alias{estimate_smooth}
\title{Describe the smooth term (for GAMs) or non-linear predictors}
\usage{
estimate_smooth(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
This function summarise the smooth term trend in terms of linear segments. Using the aproximative derivative, it separates a non-linear vector into quasi-linear segments (in which the trend is either positive or negative). Each of this segment its characterised by its beginning, end, size (in proportion, relative to the total size) trend (the linear regression coefficient) and linearity (the R2 of the linear regression).
}
\details{
See the documentation for your object's class:
\itemize{
 \item{\link[=estimate_smooth.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
