% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_smooth.R
\name{estimate_smooth.stanreg}
\alias{estimate_smooth.stanreg}
\title{Describe the smooth term (for GAMs) or non-linear predictors}
\usage{
\method{estimate_smooth}{stanreg}(model, smooth = NULL, levels = NULL,
  length = 200, transform = "response", estimate = "median", ...)
}
\arguments{
\item{model}{Bayesian model.}

\item{smooth}{A character indicating the name of the "smooth" term.}

\item{levels}{A character vectors indicating the variables over which the slope will be computed. If NULL (default), it will select all the remaining predictors.}

\item{length}{Length of numeric target variables that applies to \link{data_grid}.}

\item{transform}{Can be "none", "response", "mu", "unlink", "log". "none" (default for contrasts) will leave the values on scale of the linear predictors. "response" (default for means) will transform them on scale of the response variable. Thus for a logistic model the default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities.}

\item{estimate}{The \href{https://easystats.github.io/bayestestR/articles/2_IndicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Describe the smooth term (for GAMs) or non-linear predictors
}
\examples{
\dontrun{
library(rstanarm)
}
}
