% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_emmeans.R, R/model_emtrends.R
\name{model_emmeans}
\alias{model_emmeans}
\alias{model_emtrends}
\title{Easy Creation of 'emmeans' Objects}
\usage{
model_emmeans(
  model,
  levels = NULL,
  fixed = NULL,
  modulate = NULL,
  transform = "response",
  ...
)

model_emtrends(
  model,
  trend = NULL,
  at = NULL,
  fixed = NULL,
  levels = NULL,
  modulate = NULL,
  ...
)
}
\arguments{
\item{model}{A statistical model.}

\item{levels, modulate}{Deprecated, use \code{at} instead.}

\item{fixed}{A character vector indicating the names of the predictors to be
"fixed" (i.e., maintained), so that the estimation is made at these values.}

\item{transform}{Is passed to the \code{type} argument in
\code{emmeans::emmeans()}. See
\href{https://CRAN.R-project.org/package=emmeans/vignettes/transformations.html}{this
vignette}. Can be \code{"none"} (default for contrasts), \code{"response"}
(default for means), \code{"mu"}, \code{"unlink"}, \code{"log"}.
\code{"none"} will leave the values on scale of the linear predictors.
\code{"response"} will transform them on scale of the response variable.
Thus for a logistic model, \code{"none"} will give estimations expressed in
log-odds (probabilities on logit scale) and \code{"response"} in terms of
probabilities.}

\item{...}{Other arguments passed for instance to \code{\link{visualisation_matrix}}.}

\item{trend}{A character vector indicating the name of the numeric variable
for which to compute the slopes.}

\item{at}{The predictor variable(s) \emph{at} which to evaluate the desired effect / mean / contrasts. Other predictors of the model that are not included here will be collapsed and "averaged" over (the effect will be estimated across them).}
}
\value{
An \code{emmeans} object.
}
\description{
The \code{model_emmeans} function is a wrapper to facilitate the usage of
\code{emmeans::emmeans()} and \code{emmeans::emtrends()}, providing a
somewhat simpler and smart API to find the variables of interest.
}
\examples{
library(modelbased)

model <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)

# By default, 'levels' is set to "Species"
model_emmeans(model)

# One can estimate marginal means at several values of a 'modulate' variable
model_emmeans(model, modulate = "Petal.Width", length = 3)
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

model_emtrends(model)
model_emtrends(model, at = "Species")
model_emtrends(model, at = "Petal.Length")
model_emtrends(model, at = c("Species", "Petal.Length"))

model <- lm(Petal.Length ~ poly(Sepal.Width, 4), data = iris)
model_emtrends(model)
model_emtrends(model, at = "Sepal.Width")
}
