% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.estimate_contrasts}
\alias{print.estimate_contrasts}
\title{Printing modelbased-objects}
\usage{
\method{print}{estimate_contrasts}(
  x,
  select = getOption("modelbased_select", NULL),
  include_grid = getOption("modelbased_include_grid", FALSE),
  full_labels = TRUE,
  ...
)
}
\arguments{
\item{x}{An object returned by the different \verb{estimate_*()} functions.}

\item{select}{Determines which columns are printed and the table layout.
There are two options for this argument:
\itemize{
\item \strong{A string expression with layout pattern}

\code{select} is a string with "tokens" enclosed in braces. These tokens will be
replaced by their associated columns, where the selected columns will be
collapsed into one column. Following tokens are replaced by the related
coefficients or statistics: \code{{estimate}}, \code{{se}}, \code{{ci}} (or \code{{ci_low}} and
\code{{ci_high}}), \code{{p}}, \code{{pd}} and \code{{stars}}. The token \code{{ci}} will be replaced
by \verb{{ci_low}, {ci_high}}. Example: \code{select = "{estimate}{stars} ({ci})"}

It is possible to create multiple columns as well. A \code{|} separates values
into new cells/columns. Example: \code{select = "{estimate} ({ci})|{p}"}.
\item \strong{A string indicating a pre-defined layout}

\code{select} can be one of the following string values, to create one of the
following pre-defined column layouts:
\itemize{
\item \code{"minimal"}: Estimates, confidence intervals and numeric p-values, in two
columns. This is equivalent to \code{select = "{estimate} ({ci})|{p}"}.
\item \code{"short"}: Estimate, standard errors and numeric p-values, in two columns.
This is equivalent to \code{select = "{estimate} ({se})|{p}"}.
\item \code{"ci"}: Estimates and confidence intervals, no asterisks for p-values.
This is equivalent to \code{select = "{estimate} ({ci})"}.
\item \code{"se"}: Estimates and standard errors, no asterisks for p-values. This is
equivalent to \code{select = "{estimate} ({se})"}.
\item \code{"ci_p"}: Estimates, confidence intervals and asterisks for p-values. This
is equivalent to \code{select = "{estimate}{stars} ({ci})"}.
\item \code{"se_p"}: Estimates, standard errors and asterisks for p-values. This is
equivalent to \code{select = "{estimate}{stars} ({se})"}..
}
}

\strong{Note 1:} glue-like syntax is still experimental in the case of more complex models
(like mixed models) and may not return expected results.

\strong{Note 2:} Using \code{select} to define columns will remove all columns related
to uncertainty (standard errors, confidence intervals), test statistics, and
p-values (and similar, like \code{pd} or \code{BF} for Bayesian models), because these
are assumed to be included or intentionally excluded when using \code{select}.}

\item{include_grid}{Logical, if \code{TRUE}, the data grid is included in the
table output. Only applies to prediction-functions like \code{estimate_relation()}
or \code{estimate_link()}.}

\item{full_labels}{Logical, if \code{TRUE} (default), all labels for focal terms
are shown. If \code{FALSE}, redundant (duplicated) label rows are removed.}

\item{...}{Arguments passed to \code{insight::format_table()} or
\code{insight::export_table()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
\code{print()} method for \strong{modelbased} objects. Can be used to tweak the output
of tables.
}
\note{
Use \code{print_html()} and \code{print_md()} to create tables in HTML or
markdown format, respectively.
}
\section{Global Options to Customize Tables when Printing}{

Columns and table layout can be customized using \code{options()}:
\itemize{
\item \code{modelbased_select}: \verb{options(modelbased_select = <string>)} will set a
default value for the \code{select} argument and can be used to define a custom
default layout for printing.
\item \code{modelbased_include_grid}: \code{options(modelbased_include_grid = TRUE)} will
set a default value for the \code{include_grid} argument and can be used to
include data grids in the output by default or not.
}
}

\examples{
\dontshow{if (insight::check_if_installed("marginaleffects", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model <- lm(Petal.Length ~ Species, data = iris)
out <- estimate_means(model, "Species")

# default
print(out)

# smaller set of columns
print(out, select = "basic")
\dontshow{\}) # examplesIf}
}
