% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{standardize}
\alias{unstandardize}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{effectsize}{\code{\link[effectsize]{standardize}}, \code{\link[effectsize:standardize]{unstandardize}}}
}}

