% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO_visualisation_recipe.R
\name{visualisation_recipe}
\alias{visualisation_recipe}
\title{Visualization helper (TODO)}
\usage{
visualisation_recipe(x, ...)
}
\arguments{
\item{x}{Something.}

\item{...}{Other arguments passed to other functions.}
}
\description{
Visualization helper (TODO)
}
\examples{
library(modelbased)

# Linear Models
model <- lm(mpg ~ wt, data = mtcars)
x <- estimate_relation(model)
# visualisation_recipe(x)

x <- estimate_response(model)
# visualisation_recipe(x)
}
