% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_recipe.R
\name{visualisation_recipe}
\alias{visualisation_recipe}
\title{Prepare objects for visualisation}
\usage{
visualisation_recipe(x, ...)
}
\arguments{
\item{x}{An easystats object.}

\item{...}{Other arguments passed to other functions.}
}
\description{
This function prepares objects for visualisation by returning a list of layers with data and geoms that can be easily plotted using for instance \code{ggplot2}. See the documentation for your object's class:
\itemize{
 \item{\link[=visualisation_recipe.estimate_predicted]{modelbased} (\code{estimate_means}, \code{estimate_contrasts}, \code{estimate_slopes}, \code{estimate_predicted}, \code{estimate_grouplevel})}
}
}
