% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_recipe.R
\name{visualisation_recipe}
\alias{visualisation_recipe}
\title{Visualization helper}
\usage{
visualisation_recipe(x, ...)
}
\arguments{
\item{x}{Something.}

\item{...}{Other arguments passed to other functions.}
}
\description{
Visualization helper
}
\examples{
library(modelbased)

# Linear Models
model <- lm(mpg ~ wt, data = mtcars)
x <- estimate_relation(model)
}
